Update dependencies (#399)
* Update dependencies

* Revert mold upgrade

* Update vulkan version test

+++ libraries/gfx/vulkan/tests/version.tests.cpp
@@ -9,10 +9,10 @@ namespace morpheus::gfx::vulkan {
 
 TEST_CASE("Test construction of std::error_code object via make_error_code", "[vulkan.version]")
 {
-    constexpr std::uint8_t expectedVariant = 0;
-    constexpr std::uint8_t expectedMajor = 1;
-    constexpr std::uint8_t expectedMinor = 3;
-    constexpr std::uint8_t expectedPatch = VK_HEADER_VERSION;
+    constexpr std::uint32_t expectedVariant = 0;
+    constexpr std::uint32_t expectedMajor = 1;
+    constexpr std::uint32_t expectedMinor = 3;
+    constexpr std::uint32_t expectedPatch = VK_HEADER_VERSION;
     constexpr Version version(VK_MAKE_API_VERSION(expectedVariant, expectedMajor, expectedMinor, expectedPatch));
     STATIC_REQUIRE(version.variant() == expectedVariant);
     STATIC_REQUIRE(version.major() == expectedMajor);
